+++ platform/mac/src/server/SKKServer.mm
@@ -153,8 +153,8 @@ static void terminate(int) {
     flag = [defaults boolForKey:SKKUserDefaultKeys::enable_private_mode] == YES;
     SKKBackEnd::theInstance().EnablePrivateMode(flag);
 
-    int length = [defaults integerForKey:SKKUserDefaultKeys::minimum_completion_length];
-    SKKBackEnd::theInstance().SetMinimumCompletionLength(length);
+    NSInteger length = [defaults integerForKey:SKKUserDefaultKeys::minimum_completion_length];
+    SKKBackEnd::theInstance().SetMinimumCompletionLength((int)length);
 }
 
 - (void)reloadDictionarySet {
