+++ src/engine/backend/SKKNumericConverter.cpp
@@ -116,13 +116,13 @@ static std::string ConvertType3(const std::string& src) {
     const char* unit1[] = { "", "万", "億", "兆", "京", "垓" };
     const char* unit2[] = { "十", "百", "千" };
     std::string result;
-    unsigned int previous_size = 0;
+    unsigned long previous_size = 0;
 
     if(src.size() == 1 && src[0] == '0') {
 	return "〇";
     }
 
-    for(unsigned i = src.find_first_not_of("0"); i < src.size(); ++ i) {
+    for(unsigned long i = src.find_first_not_of("0"); i < src.size(); ++ i) {
 	switch(src[i]) {
 	case '2':
 	    result += "二";
@@ -150,7 +150,7 @@ static std::string ConvertType3(const std::string& src) {
 	    break;
 	}
 
-	int distance = src.size() - i;
+	unsigned long distance = src.size() - i;
 
 	// 「十、百、千」以外の位
 	if(distance > 4 && (distance - 1) % 4 == 0) {
@@ -194,13 +194,13 @@ static std::string ConvertType5(const std::string& src) {
     const char* unit1[] = { "", "萬", "億", "兆", "京", "垓" };
     const char* unit2[] = { "拾", "百", "阡" };
     std::string result;
-    unsigned int previous_size = 0;
+    unsigned long previous_size = 0;
 
     if(src.size() == 1 && src[0] == '0') {
 	return "零";
     }
 
-    for(unsigned i = src.find_first_not_of("0"); i < src.size(); ++ i) {
+    for(unsigned long i = src.find_first_not_of("0"); i < src.size(); ++ i) {
 	switch(src[i]) {
 	case '1':
 	    result += "壱";
@@ -231,7 +231,7 @@ static std::string ConvertType5(const std::string& src) {
 	    break;
 	}
 
-	int distance = src.size() - i;
+	unsigned long distance = src.size() - i;
 
 	// 「十、百、千」以外の位
 	if(distance > 4 && (distance - 1) % 4 == 0) {
